requirement.txt
streamlit
pydantic
pillow
opencv-python
pytesseract
pdf2image
